+++ tests/antares/delete/test_delete_api.py
@@ -31,9 +31,9 @@ from antares.craft.model.st_storage import STStorage
 from antares.craft.model.thermal import ThermalCluster
 from antares.craft.service.api_services.area_api import AreaApiService
 from antares.craft.service.api_services.binding_constraint_api import BindingConstraintApiService
-from antares.craft.service.api_services.hydro_api import HydroApiService
 from antares.craft.service.api_services.link_api import LinkApiService
 from antares.craft.service.api_services.renewable_api import RenewableApiService
+from antares.craft.service.api_services.services.hydro import HydroApiService
 from antares.craft.service.api_services.services.output import OutputApiService
 from antares.craft.service.api_services.st_storage_api import ShortTermStorageApiService
 from antares.craft.service.api_services.study_api import StudyApiService
